refactor(files): lift the scan rule into a module the ingest can share
Pinchy is about to read scanned PDFs in a second place — the knowledge-base
ingest — and two implementations of "is this page a scan, and how large may it
be rendered" would mean two different texts for the same page. A citation could
then point at a chunk the agent cannot find when it opens the file, which is
exactly the failure the attribution graders exist to detect.

So the rule and the render budget move into `pdf-scan-rule.ts`: no imports, and
none may be added, because web pulls it into `next build`. The canonical copy
has to live in the plugin even though web is the core — a plugin is deployed as
a self-contained directory and can import nothing outside itself, while web is
free to reach across. The header says so.

The render CALL deliberately does not move: web and the plugin resolve
different pdfjs majors (6.x and 5.x) whose `page.render()` parameters differ, so
a shared function would be type-checked against one and executed against the
other. Each side keeps its own ~20 lines and both read the one number that must
not drift.

`isScannedPage` stays re-exported from pdf-extract, so its existing tests keep
covering it through the path they always used.
